pyproject.toml

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "marl-cooperative-cpu-memory-optimizer"
version = "0.1.0"
description = "Multi-Agent Reinforcement Learning for Cooperative CPU and Memory Resource Optimization"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.1.0",
    "numpy>=1.24",
    "matplotlib>=3.8",
    "PyYAML>=6.0",
]
